Set-up command line application for Event Loop experimentation
It can be a bit counter-intuitive how some problems should be solved
when only working with the Revolt Event Loop and without another library
like `Amp` to take care of ensuring you use the loop correctly.

This sets up a simple Symfony command line application in which we can
add commands for certain demos that can help people understand how the
event loop can be used.