Add example that requires EventLoop::run at the end of index.php
A lot of async tasks will automatically be orchestrated on the event
loop without any additional code on Drupal's part. However, this only
works as long as the event loop is blocked on something (i.e. waiting
for a suspension). If all the code that is executed from `index.php` has
completed the process will simply exit. Which means that any code that
might still need to run (e.g. a clean-up task) will never be executed.
By adding `EventLoop::run` the process will keep running until the
scheduled async taks are executed.