Add BigPipe demo
For BigPipe we want to dispatch a set of tasks to render placeholder
values which may take different lengths of times (e.g. due to different
cache warmths; database query complexity or even communication with
external services). We implement a `stream` helper function which can
take a set of async operations and then yield the results as they become
available. This allows easily looping through the results of the
operations in calling code.